Wikipedia対応・リーダー表示の検索ボックスの対応・GIFへの緊急対応

+++ read.php
@@ -4,28 +4,57 @@
 require_once('components/xss.php');
 require_once('components/components.php');
 
-$url = $_REQUEST['url'];
+if(isset($_REQUEST['url'])){
+    $url = $_REQUEST['url'];
 
-//画像ファイル化によって対応を分ける
-if(is_image_url($url)){
-    header('Content-Type: image/webp');
-    $imagefile_path = save_with_webp($url);
-    readfile($imagefile_path);
-    unlink($imagefile_path);
+    //画像ファイル化によって対応を分ける
+    if(is_image_url($url)){
+        $root_url = $_REQUEST['root_url'];
+
+        header('Content-Type: image/webp');
+
+        $imagefile_path = save_with_webp($url, $root_url);
+        readfile($imagefile_path);
+        unlink($imagefile_path);
+    }
+    else{
+        if(isDomainUrlonList($url)){
+            if (isset($_REQUEST['extra'])){
+                $url = $url . $_REQUEST['titles'];
+            }
+
+            $content_raw_array = Readability_Raw_Extra($url);
+
+            $title = $content_raw_array[0];
+            $content_raw = $content_raw_array[1];
+
+            //HTMLのリンクを変換
+            $content_url_prefix = Add_Prefix_2_Link($content_raw, "http://localhost/sakusaku-net/read.php?url=");
+
+            //HTMLのリンクを変換
+            $content_url_prefix = Add_Prefix_2_Img_Src($content_url_prefix, "http://localhost/sakusaku-net/read.php?root_url={$url}&url=");
+        }
+        else{
+            //Rawの本文を取得
+            $content_raw_array = Readability_Raw($url);
+
+            $title = $content_raw_array[0];
+            $content_raw = $content_raw_array[1];
+
+            //HTMLのリンクを変換
+            $content_url_prefix = Add_Prefix_2_Link($content_raw, "http://localhost/sakusaku-net/read.php?url=");
+
+            //HTMLのリンクを変換
+            $content_url_prefix = Add_Prefix_2_Img_Src($content_url_prefix, "http://localhost/sakusaku-net/read.php?root_url={$url}&url=");
+        }
+    }
 }
 else{
-    //Rawの本文を取得
-    $content_raw_array = Readability_Raw($url);
-
-    $title = $content_raw_array[0];
-    $content_raw = $content_raw_array[1];
-
-    //HTMLのリンクを変換
-    $content_url_prefix = Add_Prefix_2_Link($content_raw, "http://localhost/sakusaku-net/read.php?url=");
-
-    //HTMLのリンクを変換
-    $content_url_prefix = Add_Prefix_2_Img_Src($content_url_prefix, "http://localhost/sakusaku-net/read.php?url=");
+    $title = "URLを指定してください";
+    $content_url_prefix = "URLが指定されていません。URLを指定して再度試してください。";
+    $url = "read.php";
 }
+
 ?>
 <!DOCTYPE html>
 <html lang="ja">
@@ -42,8 +71,8 @@ else{
 <body>
     <header>
         <h1><a href="index.php"></p><span style="color: red;">サクサク</span><span  class="logo_title_net">ネット</span></a></h1>
-        <form action="search.php" method="get" class="search_container">
-            <input type="text" class="search_box_input" placeholder="URL..." name="keyword">
+        <form action="read.php" method="get" class="search_container">
+            <input type="text" class="search_box_input" placeholder="URL..." name="url">
         </form>
         <a href="<?php echo $url?>">元ページを表示</a>
     </header>
